refactor: separate terminal browser responsibilities
